+++ zulu/__init__.py
@@ -12,6 +12,6 @@ from .__pkg__ import (
 )
 
 
-from .api import now, parse
+from .api import format, now, parse
 from .datetime import DateTime
-from .parser import ParseError
+from .parser import ParseError, ISO8601, TIMESTAMP